Fix error message (#4849)
If the platform target is explicitly set to x86, it sets the MSBuild architecture to the same. It then passes over the real problem when using dotnet build and returns a confusing error message instead.

ResolveComReferences target owners explicitly said they would not port it to .NET Core and others (so far) haven't either. This clarifies the error message for all such targets.