If the TFM is net472 (or any other non .NET Core TFM) but something still forces the GenerateRuntimeConfigurationFiles target to run (can be done by setting the same property explicitely to true), there will be no PlatformLibrary available, but the project will otherwise look just like self-contained app (other than the TFM).
Before #3697 the frameworks processing would simply be skipped due to the project not being framework dependent. After that change we still try to process frameworks. The fix is to check for availability of any PlatformLibrary, and if not available treat it as if there are no frameworks (just like before #3697.