Use directory-of-MSBuild in node handshake
Because CommunicationsUtilities is in Shared, it can get compiled into multiple
different assemblies. Important here are MSBuild.exe, MSBuildTaskHost.exe and
Microsoft.Build.dll.

If the method called during the handshake is from a mismatched assembly from
the host, the handshake can spuriously mismatch. But we can't unify on a
single assembly because MSBuildTaskHost doesn't load any other modern assemblies
(only .NET 3.5 MSBuild assemblies).

Instead, use the directory of the running MSBuild assembly in the handshake.
All current distributions of MSBuild require having all the assemblies in the
same directory and loading them together as a set, so this should not lose
generality.

Also added logging of the components of the hash under
MSBUILDDEBUGCOMM.

Co-authored-by: Nathan Mytelka <Forgind@users.noreply.github.com>