Normalize on Debug case for CI script
Inside the CI script, we were specifying `debug` as the configuration, but in
the YAML we copied up the `Debug` directory. That worked because somewhere in
the guts of things, test logs got emitted to `Debug`, but didn't actually
capture the build log. Specifying with a capital letter should put everything
in the same folder for happy upload.