Fixes ClickOnce problem
Fixes #4824
It fixes it by essentially forcing a rebuild of all project files if this save method is ever called. That is not optimal. An alternative solution proposed is to write a special version of Save in ProjectRootElement and have users call that instead. This version is simpler but not as efficient.