Copy in-SDK MSBuild extensions to bootstrap (#4981)
The second-stage build was intermittently failing on Linux because NuGet's
solution-restore logic wasn't being used. It wasn't being used because it
depends on being placed in `SolutionFile/ImportAfter`, but it wasn't being
copied from the last-known-good SDK to the bootstrap folder.

Fixes #4971 by copying `Current\**` (except the might-have-been-updated
`Microsoft.Common.props` which is already there) into the bootstrap folder
on Core.