Updating BootstrapUtil to look for packages in additional folders (#5074)
We want teams to be able to install bootstrapper packages from outside the VS MSIs where they currently come from. Right now we only enumerate packages from a single directory that's defined in the HKEY_LOCAL_MACHINE\SOFTWARE\Microsoft\GenericBootstrapper reg key. I'm making it so if there's any additional paths specified under the AdditionalPackagePaths sub key we'll look for bootstrapper packages in those paths as well.

One small issue with this change is that when a bootstrapper specified in a clickonce or setup project isn't found we give a warning like this:

Item '.NETFramework,Version=v4.7.2' could not be located in 'C:\Program Files (x86)\Microsoft SDKs\ClickOnce Bootstrapper'

Ideally this would be updated to also mention any possible additional paths that have been specified using this new support. But since that would require a string change and a new public interface to properly support I've decided to delay that part of the change until 16.6.