Keep track of ProjectRootElements that change via save to a stream Fixes #4824 (#5069)
Writing to a stream may or may not involve modifying a project, so we previously did not update _lastWriteTimeWhenRead to reflect changes from a stream. When we changed how we calculate MSBuildAllProjects to only include the most recently modified project (so we don't have to look through a potentially long list of irrelevant projects), we started missing out on project files that were written via writing a stream but not modified explicitly with a non-stream write—these would not be added to MSBuildAllProjects because they weren't the "most recently modified" project, causing them never to be built if the most recently written project according to MSBuildAllProjects had been written before the last build. This caused the error in ClickOnce publishing, so this fixes #4824.