Introduce ExecuteTask ETW (#5004)
This creates a new eventsource pair: ExecuteTaskStart and ExecuteTaskStop. Together, they allow a user to see the name of a task and how long it took to execute (including exact start and stop time) for each task. Another pair is added to capture how long each of yield and reacquire takes (if relevant). By subtracting the yield and reacquire times from the total time to execute the task, one can obtain the time spent actually executing the task.