Allow data columns in grids to shrink to fit the screen buffer width
Updated layout logic for grid to allow a single column to be maked as `shrinkIfNeeded`. If the screen buffer is not wide enough to fit the data, the shrinkable column will be shrunk to fit the screen exactly. If the screen is so small so that there isn't even room for that, the rendering uses a minimum width, so it looks a bit ugly, but not as bad as it might otherwise look with huge wrapped text.

The only help grid that is marked as having an optional column is the one that lists the installed templates.

Fixes these issues:
* dotnet new -l : Table view can be improved on. #361
* dotnet new usage results in line wrapping under default console size #1815
* The table in "dotnet new" is too big #2062