Integration work to support compiling R2R images using crossgen2 (#4207)
These changes enable the SDK to download and use the crossgen2 package and use crossgen2 as the ReadyToRun compiler. Unlike crossgen, crossgen2 is a tool in its own package, and no longer part of the runtime package.

The supported scenarios for crossgen2 are win-x64->win-x64 and linux-x64->linux-x64 compilations, and is only enabled if the PublishReadyToRunUseCrossgen2 property is set to true.

Also, crossgen2 currently still does not have the ability to emit native symbols for the compiled ReadyToRun images, so for those we will use crossgen from the runtime pack.