Pass MSBuild Task Id via project's ParentProjectBuildEventContext (#5013)
I've played with Mark's change #4142 and thought about it, and I think the best thing to do is to take Mark's change as is. I agree that it doesn't capture scenarios where a project's build is spawned via IBuildEngine* APIs and not via the MSBuild task, but we don't care about those as much. Even if we just cover the MSBuild task like this, it will be a huge improvement that covers most if not all scenarios that we really care about.

It's pretty low-risk and non-invasive. I've checked that in this expression:
```
issuingEntry.Request.CurrentMSBuildTask ?? issuingEntry.Request.BuildEventContext
```

all the fields on the build event context are the same, except `BuildRequestId`, `TargetId`and `TaskId`. Since `BuildRequestId` is just `GetHashCode()` I'm not worried about it. `TargetId` and `TaskId` acquire meaningful values instead of being set to -1. I've verified that the valid task id is propagated all the way to the `ParentProjectBuildEventContext` for the project being built. I've added a unit-test to validate that.

I've also made a change in the binlog viewer to take advantage of this if present. It is a major improvement that will significantly clean up the tree and get rid of most low-relevance "project builds", or rather reparent them where they belong and get them out of the way.

Fixes #3939 

Co-authored-by: Kharitonov, Mark <Mark.Kharitonov@ceridian.com>