Use new HostModel APIs (#10849)
Update GenerateBundle task and a few others to use the new APIs in HostModel intruduced by https://github.com/dotnet/runtime/pull/33413.
This change is in preparation for moving to new PublishSingleFile semantics in .net 5.

In GenerateBundle task, the SDK constructs the bundler with `BundleAllContent` option for all builds.
This keeps the behavior of PublishSingleFile unchanged until Host/Runtime components of the feature are ready.