Don't hide AfterTargets failures
`BuildResult` had code in one of its constructors specifically to handle
AfterTargets failures, making sure that they are propagated to the
referenced target. This didn't work in multi-proc builds for multiple
reasons:

1) The process orchestrating the build may have had no visibility into
targets other than those requested to be executed by the worker process.
This means that the `existingResults` argument typically didn't contain
results for the targets with AfterTargets specified, meaning it had no
way of knowing that they had failed.

2) The process orchestrating the build may not have loaded the project
file handled by the worker process. This means that the
`additionalTargetsToCheck` argument could have been `null`, again making
it impossible for `BuildResult` to detect the dependency.

This change moves the logic of calculating target failures to the worker
process, namely to `TargetBuilder`. A new flag `AfterTargetsHaveFailed`
is introduced and calculated for each requested target after all targets
have executed. This flag is then passed back as part of `TargetResult`
and trivially used in `BuildResult`'s `OverallResult`. Note that targets
with failed AfterTargets can still succeed in terms of their
`ResultCode`, i.e. the failure is kept only in form of this bolt-on bit.

The now unused `GetAfterTargetsForDefaultTargets` is deleted.

Fixes #3345