Turn off logging item metadata for task parameters.
Item metadata is logged for AddItem and RemoveItem (intrinsic tasks) as well as Output items.

Remove an overload of GetParameterText that accepts params. Everyone calling it was calling an unnecessary ToArray().
Use List instead of IList to be able to call the overload that accepts IList.

Fix a cast error in previous commit. Enumerating a non-generic IDictionary returned DictionaryEntry instead of KeyValuePair<,>. Try cast the IDictionary to IDictionary<K,V> first to get KeyValuePair<,> and if the cast fails, fall back to DictionaryEntry (this codepath shouldn't be hit).