Refactor translator APIs
ITranslator methods now take an ObjectTranslator delegate which can
translate an object in both directions, as opposed to requiring that
objects implement ITranslatable and that a factory be passed in.

The TranslatorHelpers class includes extension methods which support
the old APIs, so call sites that use ITranslatable objects and
NodePacketValueFactory still work.