Filter DTB packages by TargetFrameworkMoniker
Previously PreprocessPackageDependenciesDesignTime returned package
dependencies for all targets, and PackageRuleHandler in the project
system would filter based on TargetFrameworkMoniker.

This change applies the filtering in this task instead.

As part of this change we are now able to remove the target prefix from
the item spec of output items. This reduces string allocation, improves
the chance of pooling and reduces string manipulation by the consumer.