Fast path when evaluating string.PadLeft(int, char)
The previous fast path was only expecting the args of int and string, but was actually given int and char. So the fast path was not taken for expressions such as:

$(VersionSuffixBuildOfTheDay.PadLeft(2, $([System.Convert]::ToChar(`0`))))

Try to interpret int and char as int and string, this ensures the fast path is taken.

This fixes a first-chance MissingMethodException.