Add Task.GetParameterLoggingOptions(string)
This is a new virtual method on Task that allows task implementations to customize how and whether to log its parameters. The default implementation preserves existing behavior, which is, log everything if LogTaskInputs is true. Overrides on specific tasks customize this behavior to avoid logging item metadata where not relevant, or not log some parameters at all.

For example since the Hash task only looks at the ItemSpec, there's no point to log metadata for the input items.

The tasks were chosen by the most output. These cause the majoriy of logging for task inputs and outputs.

This does not touch tasks important for build investigations such as Rar and Csc and NuGet-related tasks. Tasks modified were carefully chosen to ensure we don't lose useful information from the logs.