Don't fail deserialization of item-spec project item metadata (#5227)
The logic to decompress large target results was refusing to rehydrate project item metadata categorized as "item-specs". In particular, the `RecursiveDir` metadatum is explicitly set by the CreateItem task and falls into this category, causing the build to fail if certain conditions are met:
1. The build is out-of-proc.
2. The number of target results moved cross-proc meets the compression threshold.

It seems unnecessary to do elaborate validation on this codepath. We implicitly trust the serialized data. Fixing the issue by relaxing the checks so that compressed item-spec metadata can be deserialized.

Fixes #5225