Allow specifying task parameter logging options via UsingTask.
This enables fine-grained control over whether:
 * to log each parameter (whether input or output)
 * or whether to log item metadata for each ITaskItem[] parameter.

When LogTaskInputs is set the default behavior is still to log all parameters and all item metadata for ITaskItem[] parameters. Since this is very verbose and hurts performance without adding any useful information it is valuable to be able to turn this logging off in certain situations.

This approach allows controlling logging via an extension to the UsingTask XML element. Even for the default AssemblyTaskFactory a ParameterGroup child element is now allowed, with subelements for parameters that can carry the new Log attribute or the new LogTask.