Add additional parameters to ResolveAssemblyReferences (#5280)
Referenceassemblyresolver target does not contain properties to control finding of extra files.  This results in extra file accesses being done when they are not wanted simply because the file does not exist.  We would like the ability to turn this functionality off without setting BuildingProject to false as that has other implications for build logic.