Eliminate boxing from OpportunisticIntern string interning paths (#5292)
This change is addressing unnecessary allocations on the string interning path. The `IInternable` implementations (except `ReuseableStringBuilder`) are struct but they end up on the managed heap anyways because of the implicit boxing taking place at the point where they are passed as an `IInternable` argument. Also, accessing the structures via the interface is relatively expensive and prevents inlining.

Fixed in this PR by genericizing the relevant methods, changing their signatures from taking `IInternable` to taking the `T` type parameter with an interface constraint.

Other changes:
- Replaced the unrolled by-character checks for hardcoded interned strings with a helper call. The old code was problematic especially given the fact that each character read was an interface call. With the changes in this PR it is not that burning but it's better to keep the generic methods small to avoid unnecessary code bloat. We let the runtime handle inlining and loop unrolling. Not to mention readability.

- Renamed `IsOrdinalEqualToStringOfSameLength` to `StartsWithStringByOrdinalComparison` and changed the semantics accordingly to support the detection of `MSBuildDummyGlobalPropertyHeader`.

- Made `InternWithStatistics` run under a lock because it's accessing non-thread-safe data structures and stat gathering (via MSBUILDDUMPOPPORTUNISTICINTERNSTATS) was failing with exceptions in parallel builds.

The net effects of this change are:
- No GC allocations in `OpportunisticIntern` steady state operation, like for example interning strings with `InternStringIfPossible()`.
- &gt;10% faster `InternableToString()` as measured with a micro-benchmark passing a mix of strings, both hardcoded (e.g. "True") and random.