Add evaluation shortcuts for commonly used String ctors (#5304)
```$([System.String]::new(%(SomeVar))``` seems to be a commonly used pattern worth optimizing. Without this change, if SomeVar is not empty, it is handled by the general late binding logic coercing the argument to ```char[]``` and calling ```String(char[])```. That is a redundant conversion from string to char array and then from char array back to string. If the argument is empty, it fails to find a matching constructor, which is technically correct but not what the user expects.

Fixes #5275