Fix ProjectRootElement.SdkReference
The logic around whether or not to mark the project dirty when changing one of the three properties for an SDK on an import was not taking into account that a new value might be coming from a setter.  The code now accepts these values and calculates whether or not the set operation affects the import element.

I added a unit test before fixing the code to verify that this is correctly verified.

Fixes #5320