Fix PublishDepsFilePath setting (#11462)
* Fix PublishDepsFilePath setting

Restore the semantics of PublishDepsFilePath to be location where the deps.json resides when published. The PublishDepsFilePath property is used by external projects (ex: asp.net) which rely on this definition.

PublishDepsFilePath is empty (by default) for PublishSingleFile, since the deps.json file is embedde within the single-file bundle.

Add a test case for this property.