Fix linker PDB behavior and add DebuggerSupport property (#12144)
* Fix linker PDB behavior

- Filter out pre-link PDBs from the publish output
- Introduce 'TrimSymbols' public option to allow removing symbols

* Keep symbols by default, add 'DebuggerSupport'

"DebuggerSupport" will now serve as the public property that controls
the linker's PDB behavior. It is intended to be used in RuntimeHostConfigurationOption
to disable debugger support as well.

When DebuggerSupport is true, the linker will output PDBs that match the
processed IL (if PDBs were present in the input).
When DebuggerSupport is false, it will remove all debug information from
the input.

Roslyn will continue to respect existing settings DebugSymbols and
DebugType, regardless of DebuggerSupport.

* PR feedback

* Introduce TrimmerRemoveSymbols

This option gets its defaults from DebuggerSupport, but can be set
explicitly.