Use a struct to pass the node handshake to allow for a more complete check (#5461)
Replaces the previous implementation of the NodeHandshake (sending a long back and forth from provider to endpoint) with sending all the relevant information as parts of a struct. This also makes it more forward-compatible, since there's an explicit version byte, and it's easier to add a field than to manipulate a long without overwriting the values of other bits. It also prevents inadvertent collisions.