Do not call GetHashCode() on enums (#5639)
Calling GetHashCode() on an enum results in boxing (i.e. a GC allocation) when running on .NET Framework. This has been addressed in .NET Core (dotnet/coreclr#7895) but we want to fix it on our side still because it's showing in Visual Studio profiles:
https://devdiv.visualstudio.com/DevDiv/_workitems/edit/1166715

This PR is also removing the calls to base.GetHashCode() which are pointless in structures that calculate the hashcode out of all their fields.