Set type to boolean in XSD where documented as such
Several property and metadata elements were documented as having boolean type, yet were not given a type in the XSD, or given string type.

This commit applies the msb:boolean type to such elements.

Doing so gives autocompletion of expected values when editing a project file in the XML editor.