Create ProjectImports.zip in memory instead of on disk (#5718)
Fixes #5383

The behavior of /bl with Embed (which is the default) is such that it will create a zip file named msbuild.ProjectImports.zip next to the binlog, and when the build completes it will embed the zip into the binlog and delete the binlog.

This poses a problem for some processes (which can be called as part of a build) that expect all files in the local directory to be accessible. Since MSBuild is holding open this ProjectImports.zip file, a process like React Native's react-native bundle command will fail the build due to sharing violations.

How tested: Validated with a simple setup of 2 csproj, the first includes the second. Verified that the resulting binlog still includes the imported projects. I changed the closing logic a little because ZipArchive requires to be closed before the data gets flushed to the stream.