Reduce ImmutableDictionary.Comparers allocations (#5744)
Fixes #5736 by keeping a prototype `ImmutableDictionary`-with-comparer
around using the `MSBuildNameIgnoreCaseComparer`, which is the primary
comparer used for all these dictionaries in MSBuild. `OrdinalIgnoreCase` is
also used and gets a special case too.

To make the type system happier, it was easiest to drop the `K` generic
parameter to the type and hard-code it to `string`--all internal non-test
uses were already doing that.