Fix SignFile task to be .NET Core aware (#5763)
SignFile looks at the .NET FX version that is passed as an argument to decide if the file specified should be SHA1 or SHA2 signed. It checks that the .NET FX version > 4.5 to turn on SHA2 signing.

.NET Core apps targeting .NET Core 3.1 pass in 3.1 as the version of the FX. This causes SignFile to use SHA1 for signing instead of SHA2.

The fix is to pass in the .NET FX identifier that the task can then compare to known identifiers for .NET FX and .NETCore. If the identifier supplied in .NETCore, we always use SHA2 signing. If the identifier supplied is .NET FX, we continue to use the existing logic.

Co-authored-by: Sujit Nayak <sujitn@microsoft.com>