Support '--' for command line options
Launching the 'dotnet build' command is possible with both '-' and '--'
switches, while launching the 'dotnet msbuild' command is only possible
with single-dash swith (for example: '-help' or '-h').
To keep both products consistent it is necessary to add the double-dash
to the list of supported switch indicators.
Changes provided in this commit implements the '--' switch indicator.

With these changes it is now possible to call 'dotnet msbuild --<switch>'
as well as 'dotnet msbuild -<switch>' and 'dotnet msbuild -<switch>',
where <switch> is any of supported command line option name.

Implementation is done by:
 - adding "--" as considered sequence when checking for correct command
   line argument's beginning,
 - considering the switch indicator's length (2 for '--', 1 for '-' and
   '\', 0 for none) when extracting the name of switch from unquoted
   command line argument