Add files missing in ClickOnce SingleFile bundle (#5795)
When a .NET 5.0 app is published with Single-File enabled, not all dependent runtime files are included in the single file bundle EXE. Some files get excluded and need to be published individually along with the SF EXE. The list of these files are output by the GenerateSingleFileBundle target in MS.NET.Publish.targets as the _FilesExcludedFromBundle output group.

ClickOnce's ResolveManifestFiles will now also be passed these additional files when we're doing a single file ClickOnce publish. This will ensure these files getting written to ClickOnce's manifest file. ClickOnce's copy files target has also been updated to copy these extra files along with the existing files (SF EXE, Launcher.exe entrypoint EXe and ClickOnce .manifest file) to the publish folder.

In addition, the GenerateLauncher ClickOnce task has been updated to do special handling for singlefilehost.exe in addition to apphost.exe. When IsSingleFile is false, apphost.exe is the name of the intermediate EXE. When IsSingleFile is true, singlefilehost.exe is the name of the intermediate EXE. If GenerateLauncher does not handle this name correctly, Launcher's entrypoint point will be set to singlefilehost.exe instead of the final application EXE.