ClickOnce deployment for self-contained apps (#5772)
ClickOnce Publish is a new feature being added to .NET Core apps. .NET Core apps have a SCD mode that ClickOnce is not aware of. ClickOnce behavior in .NET FX is to filter out references for files that below to the .NET Framework. This behavior cause ClickOnce to filter out .NET Core assemblies as well in SCD mode which will cause app launch to fail.

To fix this, the ResolveManifestFiles ClickOnce task which is responsible for filtering is being passed following additional arguments:
bool IsSelfContainedPublish
bool IsSingleFilePublish
ITaskItem[] RuntimePackAssets

The task's filtering routine will now do a lookup of the references against RuntimePackAssets and will not filter the reference if it is found in RuntimePackAssets when SCD=true and SingleFile=false.