Add new project's condition in parsing solution
When reading the solution .sln file each project is parsed by catching
the "Project(" sequence, and the parsing is stopped when reaching the
EndProject label.
However, in case of having the solution file malformed
(see: )
it is possible, that one of projects won't have it's EndProject label,
so originally it will be the only one being added to the projects list.

To avoid such situation, new condition has been added to the project's
parsing procedure, which responsibility is to check against additional
"Project(" sequence BEFORE meeting EndProject.
This situation indicates malformed solution file.

To handle this situation, additional 'else if' statement logs the
warning and recursively starts to parse another project.
When getting back from reading incorrectly nested project, the original
project's parsing procedure is stopped and whole procedure continues.