Add Project's name and it's GUID to MSB5009 error
The MSB5009 - SolutionParseNestedProjectError has been printed for a
certain project but without specifying the project's name or even GUID
which was making this issue difficult to debug.
To make it more developer/user friendly, the MSB5009 error has been
provided with the invalid project's name and GUID embedded in the error
message.
Message is extended by string arguments separate for ProjectName and
ProjectGUID, so from now on it's required to add them as a
BuildEventFileInfo constructor arguments.

NOTE: Though error message is automaticaly generated for each language,
some are still not translated.