Separate original MSB5009 error message from the extended one
The extended MSB5009 error message requires ProjectName and ProjectGuid
as arguments otherwise raw "{0}" will be displayed to the console in
case of error.
The MSB5009 error messsage is used in other places than the one covered
in the extension implementation, so each place using this error has to
be covered with the changes.

NOTE: The only exception is the SolutionFile.ParseNestedProjects() method
which uses the MSB5009 error too, but at the point of calling this error,
there are no project's name nor GUID available to display, so it should
continue to use the original error message string.
To make it possible:
 - the original error string has been kept,
 - while the extended one has been renamed to
   "SolutionParseNestedProjectErrorWithNameAndGuid"
   to indicate it's extended content and required arguments

Leaving the MSB5009 error message in the original form for
SolutionFile.ParseNestedProject() method doesn't mean that no
information will be given to developer/user.
Note that it already calls the BuildEventFileInfo constructor with full
path and current line's number, which will ease the debugging in case of
having MSB5009 error.