Allow overriding exp branch optprof sources
Before this change, the 'get optprof data from master for exp branches'
logic always overrode the parameter that allows overriding the source
branch, so if an exp branch needed data from non-master it broke.

Check to see that the passed-in value is the default (the real source
branch name) before overriding it.