Don't hang builds on BuildManager internal errors (#5917)
The logic to handle exceptions thrown on logging and communication threads was trying to make the build submission fail with the exception but it didn't always complete it. This manifested as the build hanging, for example when building with MSBUILDNODECONNECTIONTIMEOUT = 0, which is almost guaranteed to throw InternalErrorException in InstantiateNode.

This PR fixes it by making sure that all submissions are completed in OnThreadException. Unlike the original code, proper care is taken to avoid race conditions - i.e. the build result is mutated only if hasn't already been passed to the completion handler.

Fixes #5911