Add OutputPath metadata to PublishItemsOutputGroupOutputs
To resolve wrapproj logic duplication, create a new metadata
“OutputPath” on  PublishItemsOutputGroupOutputs at the end of Publish
target to represent the content of PublishDir. By the time the itemgroup
is created, the files in this itemgroup should exist on disk. So that
wapproj target do not need to prob the content of PublishDir.

Wapproj targets should depend on Publish target with additional props
being set. And it should set PublishDir to desire end output directory
to avoid an extra copying.