Fix AssemblyUtilities.CultureInfoHasGetCultures
The `Initialize()` method must be called before testing the `s_cultureInfoGetCultureMethod` static field.

This might go unnoticed if the `Initialize` method was called earlier through some other code path but the result of calling `CultureInfoHasGetCultures()` would be wrong if `Initialize()` was not called beforehand. For example, in the static constructor of `Microsoft.Build.Tasks.CultureInfoCache`.

Note: I stumbled on this when [diagnosing an issue][1] with the AssignCulture task.

[1]: https://github.com/Humanizr/Humanizer/issues/1021#issuecomment-753682963