Prevent excess node buildup
There have been several reports of nodes building up to dozens, hundreds, or even thousands of MSBuild processes. Part of the problem is the inability for certain processes, such as Visual Studio, to connect to nodes to shut them down via ShutdownAllNodes because it was taking longer to connect to them than expected. This was corrected in the previous commit by extending the timeout. This commit adds that if more nodes are in use than should be, we try to shut down all nodes not connected to the current node. This would instantly fail any other builds running concurrently, but that should be rare. This resolves #6020.