Remove unnecessary work and refactor CompareAssemblyIdentity (#5973)
Fixes #3930.

We remove the unused native call of CompareAssemblyIdentity (fusion.dll) and keep the custom managed implementation that duplicates it. Fusion implementation seems to be faster than its managed replacement, but it is only available for Windows. We, however, not sure that we run this code often enough for the difference to be noticeable. Current changes get us to a consistent implementation everywhere and are also a perf improvement since the unused native call won't happen.

* Remove unnecessary work and refactor CompareAssemblyIdentity (#3930)